fix: resolve nightly whole-project clang-tidy narrowing + main exception-escape
The nightly whole-project clang-tidy job was red since 2026-05-30 with 8
errors that slipped past the PR gate (which only runs clang-tidy-diff on
changed lines): 7 narrowing-conversions introduced by widening the
statistics counters to int64_t (#27, #28, #30) and 1 exception-escape in
main().

- statistics_serializer.cpp / statistics_manager.cpp / game_view_model_state.cpp:
  make the implicit long->double conversions explicit (static_cast<double>),
  and widen the total_completed_games accumulator to int64_t. Behaviour is
  byte-identical -- the casts only make already-occurring conversions explicit.
- main.cpp: wrap the body in a top-level try/catch that logs via
  spdlog::critical and returns EXIT_FAILURE, satisfying bugprone-exception-escape
  with a diagnostic backstop instead of std::terminate.

+++ src/core/statistics_serializer.cpp
@@ -339,7 +339,9 @@ std::expected<void, StatisticsError> exportAggregateStatsCsv(const AggregateStat
             // Calculate completion rate
             double completion_rate = 0.0;
             if (stats.games_played[i] > 0) {
-                completion_rate = (stats.games_completed[i] / static_cast<double>(stats.games_played[i])) * 100.0;
+                completion_rate =
+                    (static_cast<double>(stats.games_completed[i]) / static_cast<double>(stats.games_played[i])) *
+                    100.0;
             }
             file << completion_rate << ",";
 
@@ -361,7 +363,8 @@ std::expected<void, StatisticsError> exportAggregateStatsCsv(const AggregateStat
         file << stats.total_completed << ",";
         double overall_completion_rate = 0.0;
         if (stats.total_games > 0) {
-            overall_completion_rate = (stats.total_completed / static_cast<double>(stats.total_games)) * 100.0;
+            overall_completion_rate =
+                (static_cast<double>(stats.total_completed) / static_cast<double>(stats.total_games)) * 100.0;
         }
         file << overall_completion_rate << ",";
         file << "N/A,N/A,N/A,N/A,N/A\n";  // N/A for time and rating columns
